+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## [unreleased]
+### Features
+- use an environment for all files/tabs (#272)
+
 ## [6.12.0] (2023-03-17)
 ### Features
 - add folding range provider (#265)